+++ .env.example
@@ -1,4 +1,4 @@
-MODE=DEVELOPMENT
+MODE=development
 
 GH_PRODUCTION_CLIENT_ID=
 GH_PRODUCTION_CLIENT_SECRET=
